Revert "[Coordinate-picker-rework 1] extracted feedbackTokens to their own file (#683)"
